+++ CHANGELOG.md
@@ -17,6 +17,16 @@ Be aware that this project is still v0.y.z which means that anything can change 
 We defined for this project that while being on major version zero we mark incompatible changes with
 new minor version numbers. Please note that this is no version handling covered by `Semver`.
 
+## 0.3.2 - 2024-09-09
+
+- Performance: `Scanner` no more holds `ScannerImpl` in a `Rc<RefCell<>>` to save time during
+creation of a new `find_iter`. Instead `ScannerImpl` is now `Clone` by wrapping the match functions
+array in an `Arc`. This makes the `Scanner` usable as static global again and has the same effect
+regarding performance.
+- `Scanner::mode_name` returns a `Option<&str>` again, instead of `Option<String>` which saves an
+additional heap allocation.
+
+
 ## 0.3.1 - 2024-09-07
 
 - Add support for lots of unicode named classes like `XID_Start` and `XID_Continue` by the help of